requirements.txt
requests
discord
asyncio
python-dotenv
googlesearch-python
google-api-python-client
BeautifulSoup4
python-youtube
faapi